Add possibility to plug 'start' and 'stop' scripts to a launcher
this allows for an easy hook for starting and stopping slave machines in case these can be controlled via scripts executed by the master node - such as virtual machines or cloud instances.
The 'start' script is executed before Jenkins attempts to launch the agent on the slave, and the 'stop' script after disconnection.