Return to original plugin groupId
The plugin groupId was changed in the pom file after the release of 1.10.
That change caused the release of 1.11 and 1.12 to fail.  The last
release available from the Jenkins update center is 1.10.

There have been significant changes in the master branch of the plugin
since the release of 1.10.  Rather than risk breaking compatibility for
existing users of the 1.10 release, I've removed the incorrect change of
groupId and accepted that this plugin cannot be released without testing
those changes.

I'm not a user of the plugin.  I'm not willing to perform that testing.
If someone else wants to adopt the plugin, perform the additional testing,
and release a new version, this change is a necessary but not sufficient
step towards that release.

Compare the current code to the 1.10 tag for the details of the changes
made since the 1.10 release.

https://issues.jenkins.io/browse/JENKINS-70301 describes the
relationship between this plugin and the WMI Windows Agents plugin.
Users that want to remove the WMI Windows Agents plugin will
also need to remove this plugin or they will need to wait until
https://github.com/jenkinsci/jenkins/pull/7568 is available in Jenkins
core 2.386 or newer.